changed structure by adding 2 routes, one for the Home page as before the other one for COR finder tool
