+++ frontend/src/api/client.ts
@@ -3,7 +3,9 @@ import axios from 'axios';
 
 
 const apiClient = axios.create({
-  baseURL: window.__ENV__.VITE_API_BASE_URL,
+  // for deploy
+  // baseURL: window.__ENV__.VITE_API_BASE_URL,
+  baseURL: 'http://localhost:8000',
   headers: {
     'Content-Type': 'application/json',
   },
