+++ frontend/src/components/Header.tsx
@@ -1,29 +1,7 @@
 import { Typography, ButtonGroup, Button, Box } from '@mui/material';
 import React from 'react';
-import { useLoader } from "../contexts/LoaderContext";
-import { useMethods } from "../contexts/MethodsContext";
 
 function Header() {
-  
-  const {
-    setDataPath,
-    setImageKeyPath,
-    setRotationAnglesDataPath,
-    setUserDefinedRotationAngles,
-    setDarks,
-    setFlats,
-  } = useLoader();
-
-
-  const handleClick = () => {
-    // Predefined values for LoaderContext
-    setDataPath("/predefined/data/path");
-    setImageKeyPath("/predefined/image/key/path");
-    setRotationAnglesDataPath("/predefined/rotation/angles/path");
-    setUserDefinedRotationAngles("0", "180", "360");
-    setDarks("/predefined/darks/file", "/predefined/darks/data/path");
-    setFlats("/predefined/flats/file", "/predefined/flats/data/path");
-  };
 
   return (
     <Box
